Milestone spec
Look at all those mocks. These tests make the - resonable in my opinion
- assumption that the interface to OctoKit is stable. So, what I'm
testing is that the issues we expect to be created given the YAML file
are being sent to OctoKit::Client#create_issue correctly. What I'm not
testing is that that amounts to an eactual issue being created.